add corrected homework_3
