+++ Homework_3/Homework_3/SceneDelegate.swift
@@ -1,0 +1,15 @@
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let rootViewController = AuthorisationViewController()
+        let navigationController = UINavigationController(rootViewController: rootViewController)
+        window = UIWindow(windowScene: windowScene)
+        window?.rootViewController = navigationController
+        window?.makeKeyAndVisible()
+    }
+}
